mcp-data-platform/requirements.txt
duckdb>=1.1.0,<2.0.0
requests>=2.32.0,<3.0.0
pytest>=8.0.0,<9.0.0
mcp>=1.26.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pandas>=2.2.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
